+++ collaboredit/package.json
@@ -20,10 +20,12 @@
   "homepage": "https://github.com/anirrudh/collaborative-edit#readme",
   "dependencies": {
     "@babel/plugin-transform-runtime": "^7.12.1",
+    "@rebass/forms": "^4.0.6",
     "@types/jest": "^26.0.15",
     "@types/node": "^14.14.9",
     "@types/react": "^17.0.0",
     "@types/react-dom": "^17.0.0",
+    "@types/rebass": "^4.0.7",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-router-dom": "^5.2.0",
@@ -34,6 +36,7 @@
     "@babel/preset-env": "^7.12.7",
     "@babel/preset-react": "^7.12.7",
     "@babel/preset-typescript": "^7.12.7",
+    "@types/rebass__forms": "^4.0.4",
     "@types/webpack": "^4.41.25",
     "@types/webpack-dev-server": "^3.11.1",
     "babel-loader": "^8.2.1",
