+++ collaboredit/src/components/simpleTextEdit.tsx
@@ -1,11 +1,13 @@
+import * as React from 'react';
 import { Box } from 'rebass';
 import { Label, Textarea } from '@rebass/forms'; 
 
-const simpleEditor = () => {
+const SimpleEditor = (props) => {
 	return(
 <div>
-<Box>
-  <Label htmlFor='comment'>Comment</Label>
+<Box p={5}
+  fontSize={4}>
+  <Label htmlFor='comment'>Collaboration Text Area</Label>
   <Textarea
     id='comment'
     name='comment'
@@ -13,3 +15,5 @@ const simpleEditor = () => {
 		  </Box>
 			  </div>);
 }
+
+export default SimpleEditor;
