+++ collaboredit/webpack.config.js
@@ -25,10 +25,12 @@ module.exports = {
   output: {
     filename: 'bundle.js',
     path: path.resolve(__dirname, 'dist'),
+    publicPath: '/'
   },
     devServer: {
     contentBase: path.join(__dirname, 'dist'),
     port: 9000,
+    historyApiFallback: true,
     open: true
   },
   plugins: [
